Scale Factor

+++ src/main/java/com/mandarina/game/main/LayerDrawer.java
@@ -1,11 +1,11 @@
 package com.mandarina.game.main;
 
 public interface LayerDrawer {
-	public void drawL1(GameDrawer g, int lvlOffsetX, int lvlOffsetY);
+	public void drawL1(GameDrawer g, double lvlOffsetX, double lvlOffsetY);
 
-	public void drawL2(GameDrawer g, int lvlOffsetX, int lvlOffsetY);
+	public void drawL2(GameDrawer g, double lvlOffsetX, double lvlOffsetY);
 
-	public void drawL3(GameDrawer g, int lvlOffsetX, int lvlOffsetY);
+	public void drawL3(GameDrawer g, double lvlOffsetX, double lvlOffsetY);
 
-	public void drawL4(GameDrawer g, int lvlOffsetX, int lvlOffsetY);
+	public void drawL4(GameDrawer g, double lvlOffsetX, double lvlOffsetY);
 }
